Move out doc-links tests into a separate module
